Move ListHelper from Desktop to Core
